Naresh_code/requirements.txt
fastapi
uvicorn
streamlit
pillow
numpy
requests
easyocr
pytesseract
pdfplumber
python-docx
